refactor(music): Adopt layered module naming in music domain
