+++ src/domains/music/services/index.ts
@@ -1,0 +1,11 @@
+/**
+ * @module @domains/music/services
+ * @remarks Barrel exports for music application services that orchestrate repositories,
+ * mappers, and cache layers.
+ * @see {@link ./music-service} for music detail reads
+ * @example
+ * ```typescript
+ * import { musicService } from '@domains/music/services'
+ * ```
+ */
+export * from './music-service'
